Added onward care section to the operational faq
The onward care section of the operational faq has a section relating to onward
referrals to GP. We don't call them referrals as they are not actually referral
messages but more of a notification of sign posting. This is to ensure primacy
is not transferred to the GP which is a major concern for them.